chore(errorLogger): cleanup
